+++ extensions/jira/src/my-filters.tsx
@@ -1,11 +1,11 @@
 import { List } from "@raycast/api";
 import { useCachedPromise, useCachedState } from "@raycast/utils";
-import { useState } from "react";
+import { useMemo, useState } from "react";
 
 import { getFilters } from "./api/filters";
-import { getIssues } from "./api/issues";
 import StatusIssueList from "./components/StatusIssueList";
 import { withJiraCredentials } from "./helpers/withJiraCredentials";
+import useIssues from "./hooks/useIssues";
 
 export function MyFilters() {
   const [query, setQuery] = useState("");
@@ -15,22 +15,9 @@ export function MyFilters() {
     keepPreviousData: true,
   });
 
-  const {
-    data: issues,
-    isLoading,
-    mutate,
-  } = useCachedPromise(
-    (filterId) => {
-      const jql = filters?.find((filter) => filter.id === filterId)?.jql;
-      if (!jql) {
-        return Promise.resolve([]);
-      }
+  const jql = useMemo(() => filters?.find((filter) => filter.id === filterId)?.jql ?? "", [filters, filterId]);
 
-      return getIssues({ jql });
-    },
-    [filterId],
-    { execute: filterId !== "" }
-  );
+  const { issues, isLoading, mutate } = useIssues(jql, { execute: !!jql });
 
   const searchBarAccessory = filters ? (
     <List.Dropdown
@@ -48,7 +35,12 @@ export function MyFilters() {
   ) : null;
 
   return (
-    <StatusIssueList issues={issues} isLoading={isLoading} mutate={mutate} searchBarAccessory={searchBarAccessory} />
+    <StatusIssueList
+      issues={issues}
+      isLoading={isLoading || isLoadingFilters}
+      mutate={mutate}
+      searchBarAccessory={searchBarAccessory}
+    />
   );
 }