+++ extensions/tmux-sessioner/src/create_new_session.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Action, ActionPanel, Form, popToRoot, showToast, Toast } from "@raycast/api";
-import { creatNewSession, getAllSession } from "./sessionUtils";
+import { creatNewSession, getAllSession } from "./utils/sessionUtils";
 
 export default function CreateNewTmuxSession() {
   const [loading, setLoading] = useState<boolean>(false);